requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
decorator==4.1.2
Django==1.11.6
django-appconf==1.0.2
django-extensions==1.9.6
django-imagekit==4.0.1
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
kombu==4.1.0
olefile==0.44
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
pilkit==2.0
Pillow==4.3.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2017.2
redis==2.10.6
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
vine==1.1.4
wcwidth==0.1.7
